Move abstract parser infrastructure from servohtmlparser.rs to parse
